demo/rpi_gateway/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rpi-gateway"
version = "0.1.0"
description = "Raspberry Pi MQTT gateway for ESP32 IoT demo"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "paho-mqtt>=2.1.0",
  "pydantic>=2.7.0",
  "google-cloud-pubsub>=2.21.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
]

[project.scripts]
rpi-gateway = "gateway.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
